feat: :sparkles: Se agrega funcion para la visibilidad de los botones dependiendo del rol

+++ app/views/permisson/index.view.php
@@ -2,7 +2,13 @@
     <div class="table-header">
         <div class="tittle-table">
             <h2>Permisos</h2>
-            <button><a href="<?= URL ?>/permisson/create">nuevo</a></button>
+            <?php
+            if (isset($this->permit['Crear']) && $this->permit['Crear'] === 1) {
+            ?>
+                <button><a href="<?= URL ?>/permisson/create">nuevo</a></button>
+            <?php
+            }
+            ?>
         </div>
         <div class="input_search">
             <input type="search" placeholder="Buscar permiso">
@@ -20,11 +26,12 @@
         </thead>
         <tbody>
             <?php
+
             use Adso\libs\DateHelper;
             use Adso\libs\Helper;
 
             foreach ($data['permisos'] as $value) {
-                ?>
+            ?>
                 <tr>
                     <td>
                         <?= $value['name_permisson'] ?>
@@ -36,11 +43,21 @@
                         <?= DateHelper::shortDate($value['updated_at']) ?>
                     </td>
                     <td>
-                    <button><a href="<?= URL ?>/permisson/editar/<?= Helper::encrypt($value['id_permission']) ?>">editar</a></button>
-                    <button><a href="<?= URL ?>/permisson/delete/<?= Helper::encrypt($value['id_permission']) ?>">eliminar</a></button>
+                        <?php
+                        if (isset($this->permit['Editar']) && $this->permit['Editar'] === 1) {
+                        ?>
+                            <button><a href="<?= URL ?>/permisson/editar/<?= Helper::encrypt($value['id_permission']) ?>">editar</a></button>
+                        <?php
+                        }
+                        if (isset($this->permit['Eliminar']) && $this->permit['Eliminar'] === 1) {
+                        ?>
+                            <button><a href="<?= URL ?>/permisson/delete/<?= Helper::encrypt($value['id_permission']) ?>">eliminar</a></button>
+                        <?php
+                        }
+                        ?>
                     </td>
                 </tr>
-                <?php
+            <?php
             }
             ?>
         </tbody>
